+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - value count for each tab
+- number of available value types in entry picker (per category and in total)
 - visual marker for timezone-dependant values
 - "What's New" under Settings (references CHANGELOG.md on GitHub main branch, = this file)
 - option to expand the date value type with ISO calendar week and day of the year